ENH: Verfiy ndims>2 when accessing v1, v2, T, X [2]

+++ src/metaTube.h
@@ -174,7 +174,7 @@ class METAIO_EXPORT MetaTube : public MetaObject
 
     int   M_GetPosition(const char*, std::vector<bool> & used) const;
 
-    float M_GetFloatFromBinaryData( int post, const char * _data,
+    float M_GetFloatFromBinaryData( size_t post, const char * _data,
       size_t readSize ) const;
 
     void  M_SetFloatIntoBinaryData( float val, char * _data, int post ) const;
